+++ app/test/actions_testing.go
@@ -19,7 +19,6 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"net/url"
-	"strconv"
 )
 
 // SubmitEntryActionsBadRequest runs the method SubmitEntry of the given controller with the given parameters and payload.
@@ -838,7 +837,7 @@ func SubmitEntryDraftActionsNotImplemented(t goatest.TInterface, ctx context.Con
 // It returns the response writer so it's possible to inspect the response headers and the media type struct written to the response.
 // If ctx is nil then context.Background() is used.
 // If service is nil then a default service is created.
-func SummarizeScoreActionsBadRequest(t goatest.TInterface, ctx context.Context, service *goa.Service, ctrl app.ActionsController, contest *int, groupBy string, page int, pageSize int, sort string, task *int, user *int) (http.ResponseWriter, error) {
+func SummarizeScoreActionsBadRequest(t goatest.TInterface, ctx context.Context, service *goa.Service, ctrl app.ActionsController) (http.ResponseWriter, error) {
 	// Setup service
 	var (
 		logBuf bytes.Buffer
@@ -858,72 +857,14 @@ func SummarizeScoreActionsBadRequest(t goatest.TInterface, ctx context.Context, 
 
 	// Setup request context
 	rw := httptest.NewRecorder()
-	query := url.Values{}
-	if contest != nil {
-		sliceVal := []string{strconv.Itoa(*contest)}
-		query["contest"] = sliceVal
-	}
-	{
-		sliceVal := []string{groupBy}
-		query["groupBy"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(page)}
-		query["page"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(pageSize)}
-		query["page_size"] = sliceVal
-	}
-	{
-		sliceVal := []string{sort}
-		query["sort"] = sliceVal
-	}
-	if task != nil {
-		sliceVal := []string{strconv.Itoa(*task)}
-		query["task"] = sliceVal
-	}
-	if user != nil {
-		sliceVal := []string{strconv.Itoa(*user)}
-		query["user"] = sliceVal
-	}
 	u := &url.URL{
-		Path:     fmt.Sprintf("/sao/v1/summarize-score"),
-		RawQuery: query.Encode(),
+		Path: fmt.Sprintf("/sao/v1/summarize-score"),
 	}
 	req, err := http.NewRequest("GET", u.String(), nil)
 	if err != nil {
 		panic("invalid test " + err.Error()) // bug
 	}
 	prms := url.Values{}
-	if contest != nil {
-		sliceVal := []string{strconv.Itoa(*contest)}
-		prms["contest"] = sliceVal
-	}
-	{
-		sliceVal := []string{groupBy}
-		prms["groupBy"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(page)}
-		prms["page"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(pageSize)}
-		prms["page_size"] = sliceVal
-	}
-	{
-		sliceVal := []string{sort}
-		prms["sort"] = sliceVal
-	}
-	if task != nil {
-		sliceVal := []string{strconv.Itoa(*task)}
-		prms["task"] = sliceVal
-	}
-	if user != nil {
-		sliceVal := []string{strconv.Itoa(*user)}
-		prms["user"] = sliceVal
-	}
 	if ctx == nil {
 		ctx = context.Background()
 	}
@@ -960,129 +901,11 @@ func SummarizeScoreActionsBadRequest(t goatest.TInterface, ctx context.Context, 
 	return rw, mt
 }
 
-// SummarizeScoreActionsNotFound runs the method SummarizeScore of the given controller with the given parameters.
-// It returns the response writer so it's possible to inspect the response headers.
-// If ctx is nil then context.Background() is used.
-// If service is nil then a default service is created.
-func SummarizeScoreActionsNotFound(t goatest.TInterface, ctx context.Context, service *goa.Service, ctrl app.ActionsController, contest *int, groupBy string, page int, pageSize int, sort string, task *int, user *int) http.ResponseWriter {
-	// Setup service
-	var (
-		logBuf bytes.Buffer
-
-		respSetter goatest.ResponseSetterFunc = func(r interface{}) {}
-	)
-	if service == nil {
-		service = goatest.Service(&logBuf, respSetter)
-	} else {
-		logger := log.New(&logBuf, "", log.Ltime)
-		service.WithLogger(goa.NewLogger(logger))
-		newEncoder := func(io.Writer) goa.Encoder { return respSetter }
-		service.Encoder = goa.NewHTTPEncoder() // Make sure the code ends up using this decoder
-		service.Encoder.Register(newEncoder, "*/*")
-	}
-
-	// Setup request context
-	rw := httptest.NewRecorder()
-	query := url.Values{}
-	if contest != nil {
-		sliceVal := []string{strconv.Itoa(*contest)}
-		query["contest"] = sliceVal
-	}
-	{
-		sliceVal := []string{groupBy}
-		query["groupBy"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(page)}
-		query["page"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(pageSize)}
-		query["page_size"] = sliceVal
-	}
-	{
-		sliceVal := []string{sort}
-		query["sort"] = sliceVal
-	}
-	if task != nil {
-		sliceVal := []string{strconv.Itoa(*task)}
-		query["task"] = sliceVal
-	}
-	if user != nil {
-		sliceVal := []string{strconv.Itoa(*user)}
-		query["user"] = sliceVal
-	}
-	u := &url.URL{
-		Path:     fmt.Sprintf("/sao/v1/summarize-score"),
-		RawQuery: query.Encode(),
-	}
-	req, err := http.NewRequest("GET", u.String(), nil)
-	if err != nil {
-		panic("invalid test " + err.Error()) // bug
-	}
-	prms := url.Values{}
-	if contest != nil {
-		sliceVal := []string{strconv.Itoa(*contest)}
-		prms["contest"] = sliceVal
-	}
-	{
-		sliceVal := []string{groupBy}
-		prms["groupBy"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(page)}
-		prms["page"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(pageSize)}
-		prms["page_size"] = sliceVal
-	}
-	{
-		sliceVal := []string{sort}
-		prms["sort"] = sliceVal
-	}
-	if task != nil {
-		sliceVal := []string{strconv.Itoa(*task)}
-		prms["task"] = sliceVal
-	}
-	if user != nil {
-		sliceVal := []string{strconv.Itoa(*user)}
-		prms["user"] = sliceVal
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	goaCtx := goa.NewContext(goa.WithAction(ctx, "ActionsTest"), rw, req, prms)
-	summarizeScoreCtx, _err := app.NewSummarizeScoreActionsContext(goaCtx, req, service)
-	if _err != nil {
-		e, ok := _err.(goa.ServiceError)
-		if !ok {
-			panic("invalid test data " + _err.Error()) // bug
-		}
-		t.Errorf("unexpected parameter validation error: %+v", e)
-		return nil
-	}
-
-	// Perform action
-	_err = ctrl.SummarizeScore(summarizeScoreCtx)
-
-	// Validate response
-	if _err != nil {
-		t.Fatalf("controller returned %+v, logs:\n%s", _err, logBuf.String())
-	}
-	if rw.Code != 404 {
-		t.Errorf("invalid response status code: got %+v, expected 404", rw.Code)
-	}
-
-	// Return results
-	return rw
-}
-
 // SummarizeScoreActionsNotImplemented runs the method SummarizeScore of the given controller with the given parameters.
 // It returns the response writer so it's possible to inspect the response headers.
 // If ctx is nil then context.Background() is used.
 // If service is nil then a default service is created.
-func SummarizeScoreActionsNotImplemented(t goatest.TInterface, ctx context.Context, service *goa.Service, ctrl app.ActionsController, contest *int, groupBy string, page int, pageSize int, sort string, task *int, user *int) http.ResponseWriter {
+func SummarizeScoreActionsNotImplemented(t goatest.TInterface, ctx context.Context, service *goa.Service, ctrl app.ActionsController) http.ResponseWriter {
 	// Setup service
 	var (
 		logBuf bytes.Buffer
@@ -1101,72 +924,14 @@ func SummarizeScoreActionsNotImplemented(t goatest.TInterface, ctx context.Conte
 
 	// Setup request context
 	rw := httptest.NewRecorder()
-	query := url.Values{}
-	if contest != nil {
-		sliceVal := []string{strconv.Itoa(*contest)}
-		query["contest"] = sliceVal
-	}
-	{
-		sliceVal := []string{groupBy}
-		query["groupBy"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(page)}
-		query["page"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(pageSize)}
-		query["page_size"] = sliceVal
-	}
-	{
-		sliceVal := []string{sort}
-		query["sort"] = sliceVal
-	}
-	if task != nil {
-		sliceVal := []string{strconv.Itoa(*task)}
-		query["task"] = sliceVal
-	}
-	if user != nil {
-		sliceVal := []string{strconv.Itoa(*user)}
-		query["user"] = sliceVal
-	}
 	u := &url.URL{
-		Path:     fmt.Sprintf("/sao/v1/summarize-score"),
-		RawQuery: query.Encode(),
+		Path: fmt.Sprintf("/sao/v1/summarize-score"),
 	}
 	req, err := http.NewRequest("GET", u.String(), nil)
 	if err != nil {
 		panic("invalid test " + err.Error()) // bug
 	}
 	prms := url.Values{}
-	if contest != nil {
-		sliceVal := []string{strconv.Itoa(*contest)}
-		prms["contest"] = sliceVal
-	}
-	{
-		sliceVal := []string{groupBy}
-		prms["groupBy"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(page)}
-		prms["page"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(pageSize)}
-		prms["page_size"] = sliceVal
-	}
-	{
-		sliceVal := []string{sort}
-		prms["sort"] = sliceVal
-	}
-	if task != nil {
-		sliceVal := []string{strconv.Itoa(*task)}
-		prms["task"] = sliceVal
-	}
-	if user != nil {
-		sliceVal := []string{strconv.Itoa(*user)}
-		prms["user"] = sliceVal
-	}
 	if ctx == nil {
 		ctx = context.Background()
 	}
@@ -1194,135 +959,4 @@ func SummarizeScoreActionsNotImplemented(t goatest.TInterface, ctx context.Conte
 
 	// Return results
 	return rw
-}
-
-// SummarizeScoreActionsOK runs the method SummarizeScore of the given controller with the given parameters.
-// It returns the response writer so it's possible to inspect the response headers and the media type struct written to the response.
-// If ctx is nil then context.Background() is used.
-// If service is nil then a default service is created.
-func SummarizeScoreActionsOK(t goatest.TInterface, ctx context.Context, service *goa.Service, ctrl app.ActionsController, contest *int, groupBy string, page int, pageSize int, sort string, task *int, user *int) (http.ResponseWriter, app.ComJossemargtSaoScoreSumCollection) {
-	// Setup service
-	var (
-		logBuf bytes.Buffer
-		resp   interface{}
-
-		respSetter goatest.ResponseSetterFunc = func(r interface{}) { resp = r }
-	)
-	if service == nil {
-		service = goatest.Service(&logBuf, respSetter)
-	} else {
-		logger := log.New(&logBuf, "", log.Ltime)
-		service.WithLogger(goa.NewLogger(logger))
-		newEncoder := func(io.Writer) goa.Encoder { return respSetter }
-		service.Encoder = goa.NewHTTPEncoder() // Make sure the code ends up using this decoder
-		service.Encoder.Register(newEncoder, "*/*")
-	}
-
-	// Setup request context
-	rw := httptest.NewRecorder()
-	query := url.Values{}
-	if contest != nil {
-		sliceVal := []string{strconv.Itoa(*contest)}
-		query["contest"] = sliceVal
-	}
-	{
-		sliceVal := []string{groupBy}
-		query["groupBy"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(page)}
-		query["page"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(pageSize)}
-		query["page_size"] = sliceVal
-	}
-	{
-		sliceVal := []string{sort}
-		query["sort"] = sliceVal
-	}
-	if task != nil {
-		sliceVal := []string{strconv.Itoa(*task)}
-		query["task"] = sliceVal
-	}
-	if user != nil {
-		sliceVal := []string{strconv.Itoa(*user)}
-		query["user"] = sliceVal
-	}
-	u := &url.URL{
-		Path:     fmt.Sprintf("/sao/v1/summarize-score"),
-		RawQuery: query.Encode(),
-	}
-	req, err := http.NewRequest("GET", u.String(), nil)
-	if err != nil {
-		panic("invalid test " + err.Error()) // bug
-	}
-	prms := url.Values{}
-	if contest != nil {
-		sliceVal := []string{strconv.Itoa(*contest)}
-		prms["contest"] = sliceVal
-	}
-	{
-		sliceVal := []string{groupBy}
-		prms["groupBy"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(page)}
-		prms["page"] = sliceVal
-	}
-	{
-		sliceVal := []string{strconv.Itoa(pageSize)}
-		prms["page_size"] = sliceVal
-	}
-	{
-		sliceVal := []string{sort}
-		prms["sort"] = sliceVal
-	}
-	if task != nil {
-		sliceVal := []string{strconv.Itoa(*task)}
-		prms["task"] = sliceVal
-	}
-	if user != nil {
-		sliceVal := []string{strconv.Itoa(*user)}
-		prms["user"] = sliceVal
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	goaCtx := goa.NewContext(goa.WithAction(ctx, "ActionsTest"), rw, req, prms)
-	summarizeScoreCtx, _err := app.NewSummarizeScoreActionsContext(goaCtx, req, service)
-	if _err != nil {
-		e, ok := _err.(goa.ServiceError)
-		if !ok {
-			panic("invalid test data " + _err.Error()) // bug
-		}
-		t.Errorf("unexpected parameter validation error: %+v", e)
-		return nil, nil
-	}
-
-	// Perform action
-	_err = ctrl.SummarizeScore(summarizeScoreCtx)
-
-	// Validate response
-	if _err != nil {
-		t.Fatalf("controller returned %+v, logs:\n%s", _err, logBuf.String())
-	}
-	if rw.Code != 200 {
-		t.Errorf("invalid response status code: got %+v, expected 200", rw.Code)
-	}
-	var mt app.ComJossemargtSaoScoreSumCollection
-	if resp != nil {
-		var _ok bool
-		mt, _ok = resp.(app.ComJossemargtSaoScoreSumCollection)
-		if !_ok {
-			t.Fatalf("invalid response media: got variable of type %T, value %+v, expected instance of app.ComJossemargtSaoScoreSumCollection", resp, resp)
-		}
-		_err = mt.Validate()
-		if _err != nil {
-			t.Errorf("invalid response media type: %s", _err)
-		}
-	}
-
-	// Return results
-	return rw, mt
 }