+++ design/embeded_types.go
@@ -6,9 +6,8 @@ import (
 )
 
 const (
-	cmsAsyncOK          = "ok"
-	cmsAsyncFail        = "fail"
-	cmsAsyncUnprocessed = "unprocessed"
+	cmsAsyncOK   = "ok"
+	cmsAsyncFail = "fail"
 )
 
 var ExecutionResult = Type("ExecutionResult", func() {
@@ -31,7 +30,8 @@ var ExecutionResult = Type("ExecutionResult", func() {
 })
 
 var CompilationResult = Type("CompilationResult", func() {
-	Description("Embedded representation of an entry compilation result")
+	Description(`Embedded representation of an entry compilation result. It can be unprocessed when compilation 
+is pending, otherwise it should be either ok or fail.`)
 	Attribute("status", String, "Execution result status", func() {
 		// cms/cms/db/submission.py:300
 		Enum(cmsAsyncOK, cmsAsyncFail, cmsAsyncUnprocessed)
@@ -41,9 +41,14 @@ var CompilationResult = Type("CompilationResult", func() {
 		Default(0)
 	})
 	Attribute("stdout", String, "Compilation process' standard output", func() {
+		Example(`Done`)
 		Default("")
 	})
 	Attribute("stderr", String, "Compilation process' standard error", func() {
+		Example(`batch.cpp: In function ‘int main()’:
+batch.cpp:7:27: warning: ignoring return value of ‘int scanf(const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
+     scanf("%d %d", &a, &b);
+                           ^`)
 		Default("")
 	})
 	Attribute("time", Number, "The spent execution CPU time", func() {
@@ -62,7 +67,7 @@ var CompilationResult = Type("CompilationResult", func() {
 
 var EvaluationResult = Type("EvaluationResult", func() {
 	Description("Embedded representation of an entry evaluation result")
-	Attribute("status", String, "Execution result status", func() {
+	Attribute("status", String, "Execution result status. It only indicates if the result has been evaluated or not", func() {
 		// cms/cms/db/submission.py:348
 		Enum(cmsAsyncOK, cmsAsyncUnprocessed)
 		Default(cmsAsyncUnprocessed)
@@ -76,8 +81,8 @@ var EvaluationResult = Type("EvaluationResult", func() {
 
 var ScoreResult = Type("ScoreResult", func() {
 	Description("Embedded representation of the entry's scoring after being evaluated")
-	Attribute("taskValue", Number, "The graded value relative to the Task score", func() {
-		Example(20.00)
+	Attribute("taskValue", Number, "The graded value relative to the Task score (percent of successful runs against task's  dataset)", func() {
+		Example(80.00)
 		Default(0)
 	})
 	Attribute("contestValue", Number, "The graded value relative to the Contest score", func() {